# api/notificationservice/notificationservice_swagger.yaml
basePath: /api/notification-service
definitions:
...
  teamsdto.TeamsNotificationChannelRequest:
    properties:
      channelName:
        type: string
      description:
        type: string
      webhookUrl:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API of the Notification service
  license:
    name: AGPL-3.0-or-later
  title: Notification Service API
  version: "1.0"
paths:
  /notification-channel/Teams/check:
    post:
      consumes:
      - application/json
      description: Check if a Teams server is able to send a test message
      parameters:
      - description: Teams channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Teams server test message sent successfully
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Check Teams server
      tags:
      - Teams-channel
...
  /notification-channel/teams:
    get:
      description: List teams notification channels by type
      parameters:
      - description: Channel type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teamsdto.TeamsNotificationChannelRequest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: List Teams Channels
      tags:
      - teams-channel
    post:
      consumes:
      - application/json
      description: Create a new teams notification channel
      parameters:
      - description: Teams channel to add
        in: body
        name: TeamsChannel
        required: true
        schema:
          $ref: '#/definitions/teamsdto.TeamsNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/teamsdto.TeamsNotificationChannelRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Create Teams Channel
      tags:
      - teams-channel
  /notification-channel/teams/{id}:
    delete:
      description: Delete a teams notification channel
      parameters:
      - description: Teams channel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Delete Teams Channel
      tags:
      - teams-channel
    put:
      consumes:
      - application/json
      description: Update an existing teams notification channel
      parameters:
      - description: Teams channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Teams channel to update
        in: body
        name: TeamsChannel
        required: true
        schema:
          $ref: '#/definitions/teamsdto.TeamsNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teamsdto.TeamsNotificationChannelRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Update Teams Channel
      tags:
      - teams-channel
...
securityDefinitions:
  KeycloakAuth:
    authorizationUrl: '{{.KeycloakAuthUrl}}/realms/{{.KeycloakRealm}}/protocol/openid-connect/auth'
    flow: implicit
    type: oauth2
swagger: "2.0"